refac: Refactor find boards of circle
