+++ src/test/groovy/net/causw/application/BoardServiceTest.groovy
@@ -28,18 +28,12 @@ import javax.validation.Validator
 class BoardServiceTest extends Specification {
     private BoardPort boardPort = Mock(BoardPort.class)
     private UserPort userPort = Mock(UserPort.class)
-    private PostPort postPort = Mock(PostPort.class)
     private CirclePort circlePort = Mock(CirclePort.class)
-    private CircleMemberPort circleMemberPort = Mock(CircleMemberPort.class)
-    private CommentPort commentPort = Mock(CommentPort.class)
     private Validator validator = Validation.buildDefaultValidatorFactory().getValidator()
     private BoardService boardService = new BoardService(
             this.boardPort,
             this.userPort,
-            this.postPort,
             this.circlePort,
-            this.circleMemberPort,
-            this.commentPort,
             this.validator
     )
 
@@ -55,188 +49,6 @@ class BoardServiceTest extends Specification {
                 false,
                 null
         )
-    }
-
-    /**
-     * Test cases for board of circle find all
-     */
-    @Test
-    def "Board find by circle normal case"() {
-        given:
-        def leader = UserDomainModel.of(
-                "test",
-                "test@cau.ac.kr",
-                "test",
-                "test1234!",
-                "20210000",
-                2021,
-                Role.LEADER_CIRCLE,
-                null,
-                UserState.ACTIVE
-        )
-
-        def circle = CircleDomainModel.of(
-                "test",
-                "test",
-                null,
-                "test_description",
-                false,
-                leader
-        )
-
-        def circleMember = CircleMemberDomainModel.of(
-                "test",
-                CircleMemberStatus.MEMBER,
-                circle,
-                "test",
-                "test",
-                null,
-                null
-        )
-        this.mockBoardDomainModel.setCircle(circle)
-
-        def mockPostDomainModel = PostDomainModel.of(
-                "test post id",
-                "test post title",
-                "test post content",
-                leader,
-                false,
-                (BoardDomainModel) this.mockBoardDomainModel,
-                null,
-                null
-        )
-
-        this.userPort.findById("test") >> Optional.of(leader)
-        this.circlePort.findById("test") >> Optional.of(circle)
-        this.postPort.findLatest("test") >> Optional.of(mockPostDomainModel)
-        this.commentPort.countByPostId("test post id") >> 0
-        this.circleMemberPort.findByUserIdAndCircleId("test", "test") >> Optional.of(circleMember)
-        this.boardPort.findByCircleId("test") >> List.of(this.mockBoardDomainModel)
-
-        when:
-        def boardResponseDtoList = this.boardService.findAllByCircleId("test", "test")
-
-        then:
-        boardResponseDtoList instanceof List<BoardResponseDto>
-        with(boardResponseDtoList) {
-            get(0).getCircleId() == "test"
-        }
-    }
-
-    @Test
-    def "Board find by circle circle already deleted"() {
-        given:
-        def leader = UserDomainModel.of(
-                "test",
-                "test@cau.ac.kr",
-                "test",
-                "test1234!",
-                "20210000",
-                2021,
-                Role.LEADER_CIRCLE,
-                null,
-                UserState.ACTIVE
-        )
-
-        def circle = CircleDomainModel.of(
-                "test",
-                "test",
-                null,
-                "test_description",
-                true,
-                leader
-        )
-
-        def circleMember = CircleMemberDomainModel.of(
-                "test",
-                CircleMemberStatus.MEMBER,
-                circle,
-                "test",
-                "test",
-                null,
-                null
-        )
-        this.mockBoardDomainModel.setCircle(circle)
-
-        this.userPort.findById("test") >> Optional.of(leader)
-        this.circlePort.findById("test") >> Optional.of(circle)
-        this.circleMemberPort.findByUserIdAndCircleId("test", "test") >> Optional.of(circleMember)
-        this.boardPort.findByCircleId("test") >> List.of(this.mockBoardDomainModel)
-
-        when:
-        this.boardService.findAllByCircleId("test", "test")
-
-        then:
-        thrown(BadRequestException)
-    }
-
-    @Test
-    def "Board find by circle invalid circle member state"() {
-        given:
-        def leader = UserDomainModel.of(
-                "test",
-                "test@cau.ac.kr",
-                "test",
-                "test1234!",
-                "20210000",
-                2021,
-                Role.LEADER_CIRCLE,
-                null,
-                UserState.ACTIVE
-        )
-
-        def circle = CircleDomainModel.of(
-                "test",
-                "test",
-                null,
-                "test_description",
-                false,
-                leader
-        )
-
-        def circleMember = CircleMemberDomainModel.of(
-                "test",
-                CircleMemberStatus.MEMBER,
-                circle,
-                "test",
-                "test",
-                null,
-                null
-        )
-        this.mockBoardDomainModel.setCircle(circle)
-
-        this.userPort.findById("test") >> Optional.of(leader)
-        this.circlePort.findById("test") >> Optional.of(circle)
-        this.circleMemberPort.findByUserIdAndCircleId("test", "test") >> Optional.of(circleMember)
-        this.boardPort.findByCircleId("test") >> List.of(this.mockBoardDomainModel)
-
-        when: "Circle member is await"
-        circleMember.setStatus(CircleMemberStatus.AWAIT)
-        this.boardService.findAllByCircleId("test", "test")
-
-        then:
-        thrown(BadRequestException)
-
-        when: "Circle member is drop"
-        circleMember.setStatus(CircleMemberStatus.DROP)
-        this.boardService.findAllByCircleId("test", "test")
-
-        then:
-        thrown(UnauthorizedException)
-
-        when: "Circle member is leave"
-        circleMember.setStatus(CircleMemberStatus.LEAVE)
-        this.boardService.findAllByCircleId("test", "test")
-
-        then:
-        thrown(BadRequestException)
-
-        when: "Circle member is reject"
-        circleMember.setStatus(CircleMemberStatus.REJECT)
-        this.boardService.findAllByCircleId("test", "test")
-
-        then:
-        thrown(UnauthorizedException)
     }
 
     /**
